Moved cargo.py to utils
